+++ tomosipo/geometry/cone.py
@@ -389,11 +389,11 @@ class ConeGeometry(ProjectionGeometry):
     def project_point(self, point):
         return self.to_vec().project_point(point)
 
-    def to_box(self):
-        """Returns an oriented box representating the detector
+    def to_vol(self):
+        """Returns a volume vector geometry representing the detector
 
-        :returns: an oriented box representating the detector
+        :returns: a volume vector geometry representing the detector
         :rtype:  `VolumeVectorGeometry`
 
         """
-        return self.to_vec().to_box()
+        return self.to_vec().to_vol()