+++ tomosipo/geometry/det_vec.py
@@ -217,10 +217,10 @@ class DetectorVectorGeometry(ProjectionGeometry):
     def to_vec(self):
         return self
 
-    def to_box(self):
-        """Returns an oriented box representating the detector
+    def to_vol(self):
+        """Returns a volume vector geometry representing the detector
 
-        :returns: an oriented box representating the detector
+        :returns: a volume vector geometry representing the detector
         :rtype:  `VolumeVectorGeometry`
 
         """